Stop showing menus and buttons when user is not logged in (#69)
There is some cleanup in error handling for the project. Biggest change is how we fetch the users profile instead of putting it in a random component we now load it after getting account details from login. This means at login on the redirect page or in the assertAuth guard.



* Simplify error handling.

Move to error handling as discussed. Actions only return errors, we don't check the status code of axios results anymore since axios already throws an error on 4xx and 5xx.

* Handle error correctly

Co-authored-by: Bram Rongen <bram@thx.network>